refactor!: organize folders by component rather than technology
