+++ api/src/routes.js
@@ -1,4 +1,4 @@
-import { authRouter } from './auth.js';
+import { authRouter } from './auth/routes.js';
 
 export const attachPublicRoutes = (app) => {
   app.use('/auth', authRouter);
