--- api/src/routes/auth.js
@@ -1,8 +1,0 @@
-import express from 'express';
-import * as controllers from '#controllers/auth';
-
-export const authRouter = express.Router();
-
-authRouter.post('/login', controllers.login);
-authRouter.post('/register', controllers.register);
-authRouter.delete('/logout', controllers.logout);
