--- api/src/services/users.js
@@ -1,7 +1,0 @@
-import { prisma } from '#database';
-
-export const getUserByEmail = (email) => {
-  return prisma.user.findUnique({
-    where: { email }
-  });
-};
